Add preliminary version of q-object shortcode for pop-ups
This is built off the q-cite shortcode, and simply matches a given accession numer (`{{< q-object "77.AO.84" >}}`) with an accession in the `objects.yml` file, and outputs a hove-over pop-up with the Cat # and an image. The accession number then links to the object entry. This could be extended to include more of the data in `objects.yml` but for now, this seems sufficient as a start.