Improve q-object shortcode
It will ignore any accession numbers that match that of the object on the current page (so 77.AO.85 on the page dedicated to 77.AO.85 won’t be a link or display an image). It also ignores any accession numbers not matching one in objects.yml, and instead just returns the plain number without link or other markup.

Lastly, this will add the relevant catalogue number after the accession number, but only on the first use of that number on the page. By default this is included in parentheses, but passing a comma to to the shortcode will have it be comma separated instead.