Disable broken links
Where hyperlink citations no longer resolve, we're displaying the text of the link but not actually linking it. This is done by using HTML character entitties so that the markdown processor doesn't link them: http:&#47;&#47;

This also adds a note to the reader regarding this issue, on the About and Copyright pages.